Extract world lighting helpers
